interface update

+++ chapter_009/interface/src/main/java/ru/asemenov/filter/servlets/DeleteUser.java
@@ -6,6 +6,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 /**
@@ -24,7 +25,9 @@ public class DeleteUser extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         resp.setContentType("text/html");
-        ConnectSql.getInstance().deleteUser(req.getParameter("login"));
-        resp.sendRedirect(String.format("%s/", req.getContextPath()));
+        HttpSession session = req.getSession();
+        if (!session.getAttribute("role").equals("administrator")) {
+            ConnectSql.getInstance().deleteUser(req.getParameter("login"));
+        }
     }
 }
